Fixing bug where SearchViewController could not focus on tab when results are displayed.
Root cause: collectionView.remembersLastSelectedIndex = true.
Removing this somehow allowed the swipe up to focus on the tab bar.

In addition, I did some clean up to the way that the search controller was created.
More inline with how the rest are created. This was due to trying anything to fix, but it's cleaner
so I'm keeping it.